requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
black==23.9.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
Django==4.2.6
django-celery-beat==2.5.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
freezegun==1.2.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.1
requests==2.31.0
rpds-py==0.13.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stripe==7.0.0
tomli==2.0.1
typing-extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
